style: Apply pre-commit formatting fixes for UnifiedXGBoostTrainer
- Add missing newlines at end of files
- Fix trailing whitespace issues
- Apply automatic formatting from pre-commit hooks

+++ ml/tests/unit/test_feature_engineering.py
@@ -514,7 +514,7 @@ class TestFeatureEngineer:
         assert fe.n_features == len(config.get_feature_names())
         # Buffer has extra space for safety (n_features + 20)
         assert len(fe.feature_buffer) >= fe.n_features
-        assert fe.feature_buffer.dtype == np.float32
+        assert fe.feature_buffer.dtype == np.float64
 
     def test_feature_engineer_with_custom_config(self) -> None:
         """
@@ -630,7 +630,7 @@ class TestFeatureEngineer:
         # Check output
         assert isinstance(features, np.ndarray)
         assert len(features) == fe.n_features
-        assert features.dtype == np.float32
+        assert features.dtype == np.float64
 
     def test_calculate_features_online_with_scaler(self) -> None:
         """
